+++ tests/ui/fail_duplicate_no_display.rs
@@ -1,9 +1,0 @@
-use numbered::Numbered;
-
-#[derive(Numbered)]
-#[numbered(u8, no_display, no_display)]
-enum Mode {
-    A,
-}
-
-fn main() {}