+++ tests/ui/fail_duplicate_no_variants.rs
@@ -1,9 +1,0 @@
-use numbered::Numbered;
-
-#[derive(Numbered)]
-#[numbered(u8, no_variants, no_variants)]
-enum Mode {
-    A,
-}
-
-fn main() {}